StaticMockJVMTest merged in MockStaticMethodTest
